Move `TestHttpApiController` to `AbpServiceConvention_Tests` class.
